Fix tests

+++ micro_manager/adaptivity/global_adaptivity.py
@@ -95,7 +95,9 @@ class GlobalAdaptivityCalculator(AdaptivityCalculator):
         global_data_for_adaptivity = dict()
         for name in data_for_adaptivity.keys():
             data_as_list = self._comm.allgather(data_for_adaptivity[name])
-            global_data_for_adaptivity[name] = np.concatenate((data_as_list[:]), axis=0)
+            global_data_for_adaptivity[name] = np.concatenate(
+                (data_as_list[:]), axis=0, dtype=np.float32
+            )
 
         self._update_similarity_dists(dt, global_data_for_adaptivity)
